gpo_product/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
deprecation==2.1.0
Faker==37.4.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login==0.6.2
frozenlist==1.7.0
gotrue==2.12.2
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.5.1
packaging==25.0
pluggy==1.6.0
postgrest==1.1.1
propcache==0.3.2
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==2.5.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
websockets==15.0.1
Werkzeug==2.3.7
WTForms==3.2.1
yarl==1.20.1
python-docx==1.0.1
PyMuPDF==1.24.5
google-generativeai==0.3.1
gunicorn==21.2.0
requests==2.31.0
Pillow==10.0.1
PyPDF2==3.0.1
pdfplumber==0.10.2
pandas==2.1.4
openpyxl==3.1.2
